+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/dsl/BinaryCompatibilityValidatorConfig.kt
@@ -53,12 +53,12 @@ public class BinaryCompatibilityValidatorConfig(
 
 
     /**
-     * Whether to verify JS API. Uses compiled TypeScript definitions.
+     * Whether to verify Kotlin JS/WASM TypeScript definitions APIs.
      *
      * See [fluxo-bcv-js](https://github.com/fluxo-kt/fluxo-bcv-js) for more info.
      */
     // https://github.com/fluxo-kt/fluxo-bcv-js
-    public var jsApiChecks: Boolean = true,
+    public var tsApiChecks: Boolean = true,
 
 
     /**